refactor: extract CarImporter, TrnImporter, HisImporter
Move processCarData/processTrnData/processHisData logic into dedicated
importer classes under JsbParser/Importers/. TrnImporter owns trade-group
ID state and all transaction import helpers. JsbImportService forwards
to internal importer instances.
